Add a simple analyze step
This is related to the problem I was seeing in #6. See [#6
(comment)](https://github.com/anmonteiro/now-static-bin/pull/6#issuecomment-441272828).
I think just returning the digest of the binary might be enough for now,
given that the most important configuration option is the server port,
but if that changes in the config it will also have changed in the
binary.

Still waiting to hear more regarding the `config` object from the Zeit
team.