Add a simple analyze step (#8)
This is related to the problem I was seeing in #6. See [#6
(comment)](https://github.com/anmonteiro/now-static-bin/pull/6#issuecomment-441272828).
I think just returning the digest of the binary might be enough but let's
also include the hash of the configuration object for now, given that
the timeout could change between deployments, even if the binary 
doesn't.

Still waiting to hear more regarding the `config` object from the Zeit
team.